requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.0
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
threadpoolctl==3.6.0
